# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QUANT INDUSTRY V1 API
  description: |
    Institutional-grade quantitative trading platform API.
    
    ## Overview
    This API provides endpoints for:
    - Real-time market data and quotes
    - Trading signals and order management
    - ML/AI brain for signal generation
    - Risk management and monitoring
    - Portfolio tracking and analytics
    - Options flow and GEX analysis
    - Backtesting and strategy optimization
    
    ## Authentication
    Currently supports API key authentication via environment variables.
    Future: JWT-based auth for multi-tenant support.
    
    ## WebSocket
    Real-time data available via WebSocket at `/ws/unified`
    
  version: 1.0.0
  contact:
    name: QUANT INDUSTRY Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://localhost:8001
    description: Backend server (alternative)
...
paths:
...
  /api/kill-switch/activate:
    post:
      tags: [risk]
      summary: Activate kill switch
      operationId: activateKillSwitch
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Kill switch activated
...
  /api/backtest/run:
    post:
      tags: [backtest]
      summary: Run backtest
      operationId: runBacktest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestConfig'
      responses:
        '200':
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResult'
...
  /api/monte-carlo/run:
    post:
      tags: [backtest]
      summary: Run Monte Carlo simulation
      operationId: runMonteCarlo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                strategy:
                  type: string
                symbol:
                  type: string
                simulations:
                  type: integer
                  default: 1000
      responses:
        '200':
          description: Monte Carlo result
          content:
            application/json:
              schema:
                type: object

  # ==================== RESEARCH ====================
...
components:
  schemas:
...
    BacktestConfig:
      type: object
      required: [strategy, symbol, start_date, end_date]
      properties:
        strategy:
          type: string
        symbol:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        initial_capital:
          type: number
          default: 100000
        params:
          type: object
    BacktestResult:
      type: object
      properties:
        total_return:
          type: number
        sharpe_ratio:
          type: number
        max_drawdown:
          type: number
        win_rate:
          type: number
        trade_count:
          type: integer
        profit_factor:
          type: number
        equity_curve:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              equity:
                type: number
        trades:
          type: array
          items:
            type: object
...